Refactor and extend to version 0.2
