+++ lib/renderers/0-2.js
@@ -13,6 +13,9 @@ import {
   isValidMarkerType
 } from '../utils/tag-names';
 import kvReduce from '../utils/kv-reduce';
+import {
+  isSafeMarker
+} from '../utils/sanitization-utils';
 
 export const MOBILEDOC_VERSION = '0.2.0';
 
@@ -156,7 +159,7 @@ export default class Renderer {
       for (let j=0, m=openTypes.length; j<m; j++) {
         let markerType = this.markerTypes[openTypes[j]];
         let [tagName, attrs=[]] = markerType;
-        if (isValidMarkerType(tagName)) {
+        if (isValidMarkerType(tagName) && isSafeMarker(tagName, attrs)) {
           let lowerCaseTagName = tagName.toLowerCase();
           if (this.markupElementRenderer[lowerCaseTagName]) {
             let attrObj = attrs.reduce(kvReduce, {});
