+++ lib/renderers/0-3.js
@@ -14,7 +14,7 @@ import {
 } from '../utils/tag-names';
 import kvReduce from '../utils/kv-reduce';
 import {
-  isUnsafeUri
+  isSafeMarker
 } from '../utils/sanitization-utils';
 
 import {
@@ -44,19 +44,6 @@ function validateVersion(version) {
   if (version !== MOBILEDOC_VERSION_0_3_0 && version !== MOBILEDOC_VERSION_0_3_1) {
     throw new Error(`Unexpected Mobiledoc version "${version}"`);
   }
-}
-
-function isSafe(tagName, attrs) {
-  if (tagName === 'a') {
-    let attrObj = attrs.reduce(kvReduce, {});
-
-    let href = attrObj.href;
-    if (href.length > 0 && isUnsafeUri(href)) {
-      return false;
-    }
-  }
-
-  return true;
 }
 
 export default class Renderer {
@@ -187,7 +174,7 @@ export default class Renderer {
       for (let j=0, m=openTypes.length; j<m; j++) {
         let markerType = this.markerTypes[openTypes[j]];
         let [tagName, attrs=[]] = markerType;
-        if (isValidMarkerType(tagName) && isSafe(tagName, attrs)) {
+        if (isValidMarkerType(tagName) && isSafeMarker(tagName, attrs)) {
           let lowerCaseTagName = tagName.toLowerCase();
           if (this.markupElementRenderer[lowerCaseTagName]) {
             let attrObj = attrs.reduce(kvReduce, {});
